# Optimize PME spread charges: process 1 atom by 5 threads