… platforms/hip/src/HipKernels.cpp