The PME optimizations consists of two commit. "Copy 3 kernels from pme.cc to HIP for future optimizations" copies the kernels without modifications, so the changes of "Optimize PME spread and interpolate kernels" are better visible. The main idea is to process one atom with 5 (PME_ORDER) threads using cross-lane operations, from the commit message:
* PME_ORDER lanes process one atom;
* PME_ORDER lanes access consecutive addresses;
* No need to permute z indices with zindexTable;
* Each component (x, y, z) of thetas is calculated by one lane;

This change improves overall performance of benchmarks with PME by 4-14% (depending on the GPU and the benchmark).
I think, this optimization can be ported to CUDA but I haven't tried this yet.
@peastman, are interested in it? Can we have it as it is or should it be moved to common if it will work well on CUDA too (and what if it will not)? Currently it has kind of an opposite direction than your work of moving code from all platforms to the Common platform.