@peastman 2024-09-06
That looks really interesting.  Let's try it for CUDA and see if it's faster there too.  For OpenCL, I guess the difficulty would be that there aren't shuffles, so you would need to use more expensive shared memory operations to transfer values between threads.

What are the reasons for the C++ code changes?  Particularly changing the order of bonded and nonbonded interactions, and changing the priority of the PME stream?