@peastman 2024-09-11
I'll try profiling and see if I can figure out what's going on.