@peastman 2024-09-11
I needed to force it to regenerate the code for the python wrapper.  I have it working now.  Here's the performance I get on RTX 4080.  For the main branch:

```
Test              Precision   Constraints   H mass (amu)   dt (fs)   Ensemble   Platform   ns/day
pme               single      HBonds        1.5            4         NVT        CUDA       2050.13
apoa1pme          single      HBonds        1.5            4         NVT        CUDA       598.781
amber20-cellulose single      HBonds        1.5            4         NVT        CUDA       157.109
```

And for this PR:

```
Test              Precision   Constraints   H mass (amu)   dt (fs)   Ensemble   Platform   ns/day
pme               single      HBonds        1.5            4         NVT        CUDA       2131.45
apoa1pme          single      HBonds        1.5            4         NVT        CUDA       638.484
amber20-cellulose single      HBonds        1.5            4         NVT        CUDA       159.681
```

It's a small but measurable speedup for the first two (4% for pme, 7% for apoa1pme), but a negligible difference for the larger system (<2% speedup, which is within the noise level).  I'll profile and see if I can tell what's going on.