@peastman 2024-09-12
On the other hand, STMV gets slightly slower with the new version (by about 2%).

Profiling shows that this change makes `gridSpreadCharge()` faster (from 153 to 92 us for ApoA1), but makes `gridInterpolateForce()` slower (from 34 to 54 us for ApoA1).

Looking further into the reasons `gridInterpolateForce()` is slower, the memory throughput decreases from 46% to 25%.  The compute throughput increases from 20% to 46%, which seems like a good thing until you look at the instruction statistics and see that the new version executes almost four times as many instructions (4.4 million in the original version, 16.6 million in the new version).  It executes twice as many instructions in each cycle, but has to get through four times as many.

Compare that to `gridSpreadCharge()`.  It has a roughly 2x increase in the number of instructions (from 3.4 million to 7 million), but that's balanced by a 2x increase in compute throughput (from 5.2% to 11.5%).  In any case, this kernel is dominated by memory access rather than compute.  Looking at the memory analysis, the big difference is in the total amount of data transferred to L2: 183 MB in the original version, decreasing to 100 MB in the new version.