@peastman 2024-09-12
Here's what I get using the new version of `gridSpreadCharge()` and the old version of `gridInterpolateForce()`.

```
Test              Precision   Constraints   H mass (amu)   dt (fs)   Ensemble   Platform   ns/day
pme               single      HBonds        1.5            4         NVT        CUDA       2151.03
apoa1pme          single      HBonds        1.5            4         NVT        CUDA       658.778
amber20-cellulose single      HBonds        1.5            4         NVT        CUDA       168.298
amber20-stmv      single      HBonds        1.5            4         NVT        CUDA       51.1136
```

In all cases it's faster than either the current code or the version in this PR.  At least for NVIDIA, it looks like that's the way to go.