@peastman 2024-09-12
Here is what I get with HIP.

Current code:

```
Test              Precision   Constraints   H mass (amu)   dt (fs)   Ensemble   Platform   ns/day
pme               single      HBonds        1.5            4         NVT        HIP        762.675
apoa1pme          single      HBonds        1.5            4         NVT        HIP        201.408
amber20-cellulose single      HBonds        1.5            4         NVT        HIP        43.2503
amber20-stmv      single      HBonds        1.5            4         NVT        HIP        15.1513
```

This PR:

```
Test              Precision   Constraints   H mass (amu)   dt (fs)   Ensemble   Platform   ns/day
pme               single      HBonds        1.5            4         NVT        HIP        818.492
apoa1pme          single      HBonds        1.5            4         NVT        HIP        216.304
amber20-cellulose single      HBonds        1.5            4         NVT        HIP        47.1901
amber20-stmv      single      HBonds        1.5            4         NVT        HIP        17.1939
```

New `gridSpreadCharge()`, old `gridInterpolateForce()`:

```
Test              Precision   Constraints   H mass (amu)   dt (fs)   Ensemble   Platform   ns/day
pme               single      HBonds        1.5            4         NVT        HIP        818.197
apoa1pme          single      HBonds        1.5            4         NVT        HIP        218.901
amber20-cellulose single      HBonds        1.5            4         NVT        HIP        47.4872
amber20-stmv      single      HBonds        1.5            4         NVT        HIP        17.2838
```

The last version seems to be the fastest for HIP as well, though the difference is smaller than for CUDA.  That's with Radeon RX 5700 XT.