@peastman 2024-09-16
What do you think about just removing the changes to `gridInterpolateForce()`, since it seems to be slower in all cases?