@ex-rzr 2024-09-17
I agree, I've seen your comment with profiling results and analysis -
really interesting, some numbers are even counter-intuitive for me (I would
like to do the same on HIP using omniperf, of course, after the release).
So it should be removed, at least for now.

But unfortunately I can finalize this pull request only in a few days, as I
will return home from my vacation by the end of this week.
Can you wait with the 8.2 release till Monday?

вт, 17 сент. 2024 г., 01:04 Peter Eastman ***@***.***>:

> What do you think about just removing the changes to
> gridInterpolateForce(), since it seems to be slower in all cases?
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/openmm/openmm/pull/4633#issuecomment-2353809449>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AABRZP4R4UV5TUUTIIOL2OLZW42VHAVCNFSM6AAAAABNIVQLUWVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDGNJTHAYDSNBUHE>
> .
> You are receiving this because you modified the open/close state.Message
> ID: ***@***.***>
>