@peastman 2024-09-23
Here are the results on an Apple M1 Pro.  It also gets a speedup.

Current code:

```
Test              Precision   Constraints   H mass (amu)   dt (fs)   Ensemble   Platform   ns/day
pme               single      HBonds        1.5            4         NVT        OpenCL     156.729
apoa1pme          single      HBonds        1.5            4         NVT        OpenCL     38.9667
amber20-cellulose single      HBonds        1.5            4         NVT        OpenCL     9.59166
amber20-stmv      single      HBonds        1.5            4         NVT        OpenCL     3.46364
```

This PR:

```
Test              Precision   Constraints   H mass (amu)   dt (fs)   Ensemble   Platform   ns/day
pme               single      HBonds        1.5            4         NVT        OpenCL     179.783
apoa1pme          single      HBonds        1.5            4         NVT        OpenCL     45.0727
amber20-cellulose single      HBonds        1.5            4         NVT        OpenCL     10.7347
amber20-stmv      single      HBonds        1.5            4         NVT        OpenCL     3.8491
```

If you're happy with the code, mark this PR as no longer a draft and I'll merge it.